added for each component its own file for data and also data type
