+++ src/app/home/component/aboutUs.tsx
@@ -1,46 +1,8 @@
 import Image from "next/image";
-import bagIcon from "public/template_content_icons/bag_icon.png";
-import babyCart from "public/template_content_icons/baby_car_icon.png";
-import worldIcon from "public/template_content_icons/world_icon.png";
-import ringIcon from "public/template_content_icons/ring_icon.png";
+import { contentTempTypes } from "@/shared/types/models";
+
 import tempOrangeStrip from "public/template_content_icons/temp_orange_strip.png";
-
-type contentTempTypes = {
-  title: string;
-  description: string;
-  icon: any;
-  color: string;
-};
-
-const contentTempData: contentTempTypes[] = [
-  {
-    title: "Morbi Etos",
-    description:
-      " Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor laudantium",
-    icon: bagIcon,
-    color: "#fe6501",
-  },
-  {
-    title: "Congue Gravida",
-    description: " Lorem ipsum dolor sit amet consectetur adipisicing elit.",
-    icon: babyCart,
-    color: "rgb(86,80,159)",
-  },
-  {
-    title: "Maecenas Node",
-    description:
-      " Lorem ipsum dolor sit amet consectetur adipisicing elit. dolor sit amet consectetur ",
-    icon: worldIcon,
-    color: "#fe6501",
-  },
-  {
-    title: "Praesent Morbi",
-    description:
-      "  amet consectetur adipisicing elit. dolor sit amet consectetur Lorem ipsum dolor sit ",
-    icon: ringIcon,
-    color: "rgb(86,80,159)",
-  },
-];
+import contentTempData from "@/shared/data/AboutUsData";
 
 export default function AboutUs() {
   return (
@@ -55,7 +17,7 @@ export default function AboutUs() {
       </div>
       <Image src={tempOrangeStrip} alt="" />
       <div className="mt-[66px] flex w-[80%] flex-row  flex-wrap items-center justify-center gap-[33px] text-center ">
-        {contentTempData.map((data: contentTempTypes, index) => {
+        {contentTempData.map((data: contentTempTypes, index: number) => {
           return (
             <div
               key={index}
