+++ src/app/shared/data/OverViewPageData.tsx
@@ -1,5 +1,6 @@
-import { LandingPageTypes } from "@/shared/types/models";
-const LandingPageData: LandingPageTypes[] = [
+import { OverViewPageTypes } from "@/shared/types/models";
+
+const OverViewPageData: OverViewPageTypes[] = [
   {
     id: 1,
     img: `https://via.placeholder.com/1980x1080`,
@@ -14,4 +15,4 @@ const LandingPageData: LandingPageTypes[] = [
   },
 ];
 
-export default LandingPageData;
+export default OverViewPageData;
