+++ src/app/shared/types/AboutUsDataTypes.tsx
@@ -1,0 +1,6 @@
+export interface contentTempTypes {
+  title: string;
+  description: string;
+  icon: any;
+  color: string;
+}
