+++ src/app/shared/types/OverViewPageTypes.tsx
@@ -1,0 +1,4 @@
+export interface OverViewPageTypes {
+  id: number;
+  img: string;
+}
