+++ src/app/shared/types/StaticsComponentDataTypes.tsx
@@ -1,0 +1,6 @@
+export interface StaticsComponentDataTypes {
+  title: string;
+  description: string;
+  number: number;
+  id: number;
+}
