+++ src/app/shared/types/classesTypes.tsx
@@ -1,0 +1,9 @@
+export type classesTypes = {
+  title: string;
+  className: string;
+  monthOld: number[];
+  monthOldText: string;
+  classSize: number;
+  classSizeText: string;
+  classImage: string;
+};
