+++ src/app/shared/types/models.ts
@@ -1,1 +1,5 @@
 export * from "./LandingPageTypes";
+export * from "./AboutUsDataTypes";
+export * from "./OverViewPageTypes";
+export * from "./StaticsComponentDataTypes";
+export * from "./classesTypes";
